+++ main.tf
@@ -25,7 +25,7 @@ locals {
           to_port     = interface.port
         }]
       }
-    ] if lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles)
+    ] if(lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) || lookup(local.apic, "auto_generate_access_leaf_switch_interface_profiles", local.defaults.apic.auto_generate_access_leaf_switch_interface_profiles))
   ])
 
   sub_interface_selectors = flatten([
@@ -48,7 +48,7 @@ locals {
           to_sub_port   = sub.port
         }]
       }
-    ] if lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles)
+    ] if(lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) || lookup(local.apic, "auto_generate_access_leaf_switch_interface_profiles", local.defaults.apic.auto_generate_access_leaf_switch_interface_profiles))
   ])
 }
 
@@ -56,7 +56,7 @@ module "aci_access_fex_interface_profile_auto" {
   source  = "netascode/access-fex-interface-profile/aci"
   version = "0.1.0"
 
-  for_each = { for fex in lookup(local.node, "fexes", {}) : fex.id => fex if lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) && local.node_role == "leaf" && lookup(local.modules, "aci_access_fex_interface_profile", true) }
+  for_each = { for fex in lookup(local.node, "fexes", {}) : fex.id => fex if(lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) || lookup(local.apic, "auto_generate_access_leaf_switch_interface_profiles", local.defaults.apic.auto_generate_access_leaf_switch_interface_profiles)) && local.node_role == "leaf" && lookup(local.modules, "aci_access_fex_interface_profile", true) }
   name     = replace("${local.node_id}:${local.node_name}:${each.value.id}", "/^(?P<id>.+):(?P<name>.+):(?P<fex>.+)$/", replace(replace(replace(lookup(local.access_policies, "fex_profile_name", local.defaults.apic.access_policies.fex_profile_name), "\\g<id>", "$id"), "\\g<name>", "$name"), "\\g<fex>", "$fex"))
 }
 
@@ -64,7 +64,7 @@ module "aci_access_leaf_interface_selector_auto" {
   source  = "netascode/access-leaf-interface-selector/aci"
   version = "0.2.1"
 
-  for_each              = { for int in local.node.interfaces : int.port => int if lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) && local.node_role == "leaf" && lookup(local.modules, "aci_access_leaf_interface_selector", true) }
+  for_each              = { for int in local.node.interfaces : int.port => int if(lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) || lookup(local.apic, "auto_generate_access_leaf_switch_interface_profiles", local.defaults.apic.auto_generate_access_leaf_switch_interface_profiles)) && local.node_role == "leaf" && lookup(local.modules, "aci_access_leaf_interface_selector", true) }
   name                  = replace(format("%s:%s", lookup(each.value, "module", local.defaults.apic.access_policies.leaf_interface_profiles.selectors.port_blocks.from_module), each.value.port), "/^(?P<mod>.+):(?P<port>.+)$/", replace(replace(lookup(local.access_policies, "leaf_interface_selector_name", local.defaults.apic.access_policies.leaf_interface_selector_name), "\\g<mod>", "$mod"), "\\g<port>", "$port"))
   interface_profile     = replace("${local.node_id}:${local.node_name}", "/^(?P<id>.+):(?P<name>.+)$/", replace(replace(lookup(local.access_policies, "leaf_interface_profile_name", local.defaults.apic.access_policies.leaf_interface_profile_name), "\\g<id>", "$id"), "\\g<name>", "$name"))
   fex_id                = lookup(each.value, "fex_id", 0)
@@ -115,7 +115,7 @@ module "aci_access_spine_interface_selector_auto" {
   source  = "netascode/access-spine-interface-selector/aci"
   version = "0.2.0"
 
-  for_each          = { for int in local.node.interfaces : int.port => int if lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) && local.node_role == "spine" && lookup(local.modules, "aci_access_spine_interface_selector", true) }
+  for_each          = { for int in local.node.interfaces : int.port => int if(lookup(local.apic, "auto_generate_switch_pod_profiles", local.defaults.apic.auto_generate_switch_pod_profiles) || lookup(local.apic, "auto_generate_access_spine_switch_interface_profiles", local.defaults.apic.auto_generate_access_spine_switch_interface_profiles)) && local.node_role == "spine" && lookup(local.modules, "aci_access_spine_interface_selector", true) }
   name              = replace(format("%s:%s", lookup(each.value, "module", local.defaults.apic.access_policies.spine_interface_profiles.selectors.port_blocks.from_module), each.value.port), "/^(?P<mod>.+):(?P<port>.+)$/", replace(replace(lookup(local.access_policies, "spine_interface_selector_name", local.defaults.apic.access_policies.spine_interface_selector_name), "\\g<mod>", "$mod"), "\\g<port>", "$port"))
   interface_profile = replace("${local.node_id}:${local.node_name}", "/^(?P<id>.+):(?P<name>.+)$/", replace(replace(lookup(local.access_policies, "spine_interface_profile_name", local.defaults.apic.access_policies.spine_interface_profile_name), "\\g<id>", "$id"), "\\g<name>", "$name"))
   policy_group      = lookup(each.value, "policy_group", null) != null ? "${each.value.policy_group}${local.defaults.apic.access_policies.spine_interface_policy_groups.name_suffix}" : ""