+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.2 (unreleased)
 
 - Add colon to allowed characters of leaf interface selector names
+- Add support for `auto_generate_access_leaf_switch_interface_profiles` and `auto_generate_access_spine_switch_interface_profiles` flags
 
 ## 0.3.1